encoding/json: add UseAllKeys() Decoder option to reject unknown keys
It is currently not possible to perform strict input validation
in JSON unmarshaling. By design unrecognized keys are silently
ignored. This patch adds an option to the Decoder that will
collect all keys that do not have a matching struct field.
When this option is enabled the Unmarshal() function emits
an error when unused keys are detected.

The encoding/json unit tests have been refactored such that
each test case is run twice: once with UseAllKeys() disabled
and once with UseAllKeys() enabled. The majority of test
cases produce identical output under the two scenarios.
For the handful of test cases that produce different output,
the expected output or error has been updated in those test cases.